add custom hook for landingNodeActions
